move !botinfo to subcommand
